add celery and redis solution for background tasks
Previous approach was not working correctly because background tasks
activated when FastAPI restarted. Probably something to do with the
event loop, which I could't debug.

FastAPI background tasks also did not work, had the same problem. I
suppose they use asyncio which is the source of the problem.

This commit installs redis in the fastapi image and has celery and
fastapi running in parallel.